chore(random): move rand sequence gen out of clientServer resource
