+++ pkg/random/random.go
@@ -1,0 +1,34 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright (C) 2024 The Falco Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package random
+
+import (
+	"crypto/rand"
+	"math/big"
+)
+
+var letters = []rune("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ")
+
+// Seq generates a random sequence of length n.
+func Seq(n int) string {
+	b := make([]rune, n)
+	lettersLen := len(letters)
+	for i := range b {
+		letterIndex, _ := rand.Int(rand.Reader, big.NewInt(int64(lettersLen)))
+		b[i] = letters[letterIndex.Int64()]
+	}
+	return string(b)
+}
